Code review: naming, final, typo, etc...

+++ core-common/src/main/java/eu/arrowhead/common/dto/ServiceDefinitionsListResponseDTO.java
@@ -11,7 +11,7 @@ public class ServiceDefinitionsListResponseDTO implements Serializable {
 	// members
 	
 	private List<ServiceDefinitionResponseDTO> data;
-	private int count;
+	private long count;
 	
 	//=================================================================================================
 	// methods
@@ -20,16 +20,16 @@ public class ServiceDefinitionsListResponseDTO implements Serializable {
 	public ServiceDefinitionsListResponseDTO() {}
 
 	//-------------------------------------------------------------------------------------------------
-	public ServiceDefinitionsListResponseDTO(final List<ServiceDefinitionResponseDTO> data, final int count) {
+	public ServiceDefinitionsListResponseDTO(final List<ServiceDefinitionResponseDTO> data, final long count) {
 		this.data = data;
 		this.count = count;
 	}
 
 	//-------------------------------------------------------------------------------------------------
 	public List<ServiceDefinitionResponseDTO> getData() { return data; }
-	public int getCount() { return count; }
+	public long getCount() { return count; }
 	
 	//-------------------------------------------------------------------------------------------------
 	public void setData(final List<ServiceDefinitionResponseDTO> data) { this.data = data; }
-	public void setCount(final int count) { this.count = count; }
+	public void setCount(final long count) { this.count = count; }
 }
